move: inner methods to lib dir and rename
